refactor logic for fetching open positions and available nfts || render open positions by default | todo; write logic regarding new position button - connect borrow module right hand side
