+++ components/AssetRow.tsx
@@ -15,11 +15,10 @@ export interface AssetRowType {
     totalBorrowed: number,
     interest: number,
     available: number,
-    positions: number,
+    positions: number
 }
-
 interface AssetRowProps {
-    data: AssetRowType,
+    data: AssetRowType
 }
 
 const AssetRow = ({ data }: AssetRowProps) => {
